feat(86baefhe9): update song api

+++ Modules/Production/app/Services/EntertainmentService.php
@@ -16,6 +16,7 @@ use Modules\Production\Repository\ProjectSongRepository;
 class EntertainmentService
 {
     const SONG_LIST_CACHE_KEY = 'project_song_list';
+
     const TASK_LIST = 'task_list';
 
     public function __construct(
@@ -26,13 +27,10 @@ class EntertainmentService
 
     /**
      * Get song list cache identifier
-     *
-     * @param string $projectUid
-     * @return string
      */
     public function getSongListCacheKey(string $projectUid): string
     {
-        return self::SONG_LIST_CACHE_KEY . "_{$projectUid}";
+        return self::SONG_LIST_CACHE_KEY."_{$projectUid}";
     }
 
     protected function resetSongListCache(string $projectUid)
@@ -46,31 +44,22 @@ class EntertainmentService
 
         if ($cache) {
             // update song list
-            
+
         }
     }
 
     /**
      * Get task cache identifier
-     *
-     * @param string $projectUid
-     * @return string
      */
     public function getTaskCacheKey(string $projectUid): string
     {
-        return self::TASK_LIST . "_{$projectUid}";
+        return self::TASK_LIST."_{$projectUid}";
     }
 
-    public function storeTask()
-    {
-
-    }
+    public function storeTask() {}
 
     /**
      * Format project song list to serve API response
-     *
-     * @param ProjectSong $data
-     * @return array
      */
     protected function formatSongList(ProjectSong $data): array
     {
@@ -81,8 +70,8 @@ class EntertainmentService
                 uid: $songList->uid,
                 name: $songList->song_name,
                 group: $data->group_name,
-                status: !$songList->latestTask ? __('global.unassigned') : __('global.assigned'),
-                status_color: !$songList->latestTask ? 'grey' : 'green'
+                status: ! $songList->latestTask ? __('global.unassigned') : __('global.assigned'),
+                status_color: ! $songList->latestTask ? 'grey' : 'green'
             );
         }
 
@@ -91,18 +80,15 @@ class EntertainmentService
 
     /**
      * Fetch main data for project song list
-     *
-     * @param integer $projectId
-     * @return ProjectSong
      */
-    protected function fetchSongList(int $projectId): ProjectSong
+    protected function fetchSongList(int $projectId): ?ProjectSong
     {
         return $this->projectSongRepo->show(
             uid: '',
             select: 'id,group_name,uid,project_id',
             relation: [
                 'items:id,song_name,project_song_id,uid',
-                'items.latestTask:entertainment_task_song_items.id,entertainment_task_song_items.song_item_id'
+                'items.latestTask:entertainment_task_song_items.id,entertainment_task_song_items.song_item_id',
             ],
             where: "project_id = {$projectId}"
         );
@@ -110,22 +96,22 @@ class EntertainmentService
 
     /**
      * Cached the song list
-     *
-     * @param string $projectUid
-     * @return array
      */
     protected function cacheSongList(string $projectUid): array
     {
-        $projectId = getIdFromUid($projectUid, new Project());
+        $projectId = getIdFromUid($projectUid, new Project);
 
         return Cache::remember(
-            key: self::SONG_LIST_CACHE_KEY . "_{$projectUid}",
+            key: self::SONG_LIST_CACHE_KEY."_{$projectUid}",
             ttl: now()->addHours(2),
             callback: function () use ($projectId) {
                 $data = $this->fetchSongList($projectId);
-        
+
                 /** @var SongListData[] */
-                $output = $this->formatSongList(data: $data);
+                $output = [];
+                if ($data) {
+                    $output = $this->formatSongList(data: $data);
+                }
 
                 return $output;
             }
@@ -134,9 +120,6 @@ class EntertainmentService
 
     /**
      * Get list of songs
-     *
-     * @param string $projectUid
-     * @return array
      */
     public function list(string $projectUid): array
     {
@@ -164,16 +147,12 @@ class EntertainmentService
 
     /**
      * Create new song group
-     *
-     * @param CreateSongData $payload
-     * @param string $projectUid
-     * @return array
      */
     public function createSong(CreateSongData $payload, string $projectUid): array
     {
         DB::beginTransaction();
         try {
-            $projectId = getIdFromUid($projectUid, new Project());
+            $projectId = getIdFromUid($projectUid, new Project);
 
             $actor = $this->generalService->me();
 
@@ -189,7 +168,7 @@ class EntertainmentService
                     $group = $this->projectSongRepo->store([
                         'project_id' => $projectId,
                         'group_name' => $item->name,
-                        'created_by' => $actor->id
+                        'created_by' => $actor->id,
                     ]);
                 }
 
@@ -209,25 +188,69 @@ class EntertainmentService
             );
         } catch (\Throwable $th) {
             DB::rollBack();
+
             return errorResponse($th);
         }
     }
 
+    /**
+     * Update song name in song list cache
+     */
+    protected function updateSongNameInCache(string $songUid, string $projectUid, string $updatedName): void
+    {
+        $cache = Cache::get($this->getSongListCacheKey($projectUid));
+
+        if ($cache) {
+            $updated = collect($cache)->map(function ($song) use ($songUid, $updatedName) {
+                if ($song->uid === $songUid) {
+                    $song->name = $updatedName;
+                }
+
+                return $song;
+            })->all();
+
+            Cache::set($this->getSongListCacheKey($projectUid), $updated, now()->addHours(2));
+        }
+    }
+
+    protected function deleteSongFromCache(string $projectUid, string $songUid): void
+    {
+        $cache = Cache::get($this->getSongListCacheKey($projectUid));
+
+        if ($cache) {
+            $find = collect($cache)->search(function ($item) use ($songUid) {
+                return $item->uid === $songUid;
+            });
+
+            if ($find !== false) {
+                $updated = collect($cache)->forget($find);
+                Cache::set($this->getSongListCacheKey($projectUid), $updated->values()->all(), now()->addHours(2));
+            }
+        }
+    }
+
+    /**
+     * Update selected songe
+     */
     public function updateSong(UpdateSongData $payload, string $projectUid, string $songUid): array
     {
         try {
-            $projectId = getIdFromUid($projectUid, new Project());
-
             $this->projectSongItemRepo->update(
                 data: [
                     'song_name' => $payload->song,
                 ],
-                id: '',
-                where: "project_id = {$projectId} AND uid = $songUid"
+                id: $songUid,
+            );
+
+            // Update cache name
+            $this->updateSongNameInCache(
+                $songUid,
+                $projectUid,
+                $payload->song
             );
 
             return generalResponse(
-                message: "Success"
+                message: 'Success update song'
             );
         } catch (\Throwable $th) {
             return errorResponse($th);
@@ -236,10 +259,6 @@ class EntertainmentService
 
     /**
      * Delete selected song
-     *
-     * @param string $projectUid
-     * @param string $songUid
-     * @return array
      */
     public function deleteSong(string $projectUid, string $songUid): array
     {
@@ -248,22 +267,26 @@ class EntertainmentService
                 uid: $songUid,
                 select: 'id',
                 relation: [
-                    'latestTask'
+                    'latestTask',
                 ]
             );
 
+            if (! $song) {
+                return errorResponse(message: 'Song not found');
+            }
+
             if ($song->latestTask) {
-                return errorResponse(message: "Failed to delete song. Please delete task related with this song.");
+                return errorResponse(message: 'Failed to delete song. Please delete task related with this song.');
             }
 
             // No task here
             $song->delete();
 
-            // Clear cache
-            $this->resetSongListCache($projectUid);
+            // Delete song from cache list
+            $this->deleteSongFromCache($projectUid, $songUid);
 
             return generalResponse(
-                message: "Success delete song list"
+                message: 'Success delete song list'
             );
         } catch (\Throwable $th) {
             return errorResponse($th);
